updated websocket dependency handling and fixed issues with guards
