+++ tests/kwargs/test_validations.py
@@ -76,12 +76,12 @@ def test_raises_exception_when_keys_are_ambiguous(handler):
 def test_raises_when_reserved_kwargs_are_misused(reserved_kwarg):
     decorator = post if reserved_kwarg != "socket" else websocket
 
-    exec(f"def test_fn({reserved_kwarg}: int) -> None: pass")
+    exec(f"async def test_fn({reserved_kwarg}: int) -> None: pass")
     handler_with_path_param = decorator("/{" + reserved_kwarg + ":int}")(locals()["test_fn"])
     with pytest.raises(ImproperlyConfiguredException):
         Starlite(route_handlers=[handler_with_path_param])
 
-    exec(f"def test_fn({reserved_kwarg}: int) -> None: pass")
+    exec(f"async def test_fn({reserved_kwarg}: int) -> None: pass")
     handler_with_dependency = decorator("/", dependencies={reserved_kwarg: Provide(my_dependency)})(locals()["test_fn"])
     with pytest.raises(ImproperlyConfiguredException):
         Starlite(route_handlers=[handler_with_dependency])
@@ -89,7 +89,7 @@ def test_raises_when_reserved_kwargs_are_misused(reserved_kwarg):
     # these kwargs are set to Any when the signature model is generated,
     # because pydantic can't handle generics for non pydantic classes. So these tests wont work for aliased parameters.
     if reserved_kwarg not in ["socket", "request"]:
-        exec(f"def test_fn({reserved_kwarg}: int = Parameter(query='my_param')) -> None: pass")
+        exec(f"async def test_fn({reserved_kwarg}: int = Parameter(query='my_param')) -> None: pass")
         handler_with_aliased_param = decorator("/")(locals()["test_fn"])
         with pytest.raises(ImproperlyConfiguredException):
             Starlite(route_handlers=[handler_with_aliased_param])
