+++ tests/route/test_validations.py
@@ -1,6 +1,6 @@
 import pytest
 
-from starlite import get, post
+from starlite import ImproperlyConfiguredException, Starlite, get, post
 from starlite.exceptions import MethodNotAllowedException
 from starlite.routing import HTTPRoute
 
@@ -21,3 +21,16 @@ async def test_http_route_raises_for_unsupported_method():
 
     with pytest.raises(MethodNotAllowedException):
         await route.handle(scope={"method": "DELETE"}, receive=lambda x: x, send=lambda x: x)
+
+
+def test_register_validation_duplicate_handlers_for_same_route_and_method():
+    @get(path="/first")
+    def first_route_handler() -> None:
+        pass
+
+    @get(path="/first")
+    def second_route_handler() -> None:
+        pass
+
+    with pytest.raises(ImproperlyConfiguredException):
+        Starlite(route_handlers=[first_route_handler, second_route_handler])
